refactor: input checkbox
