--- src/components/Input/inputCheckbox/styled/checkboxContainer.js
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const CheckboxContainer = styled.div`
-    display: inline-block;
-`;
-
-export default CheckboxContainer;
