--- src/components/Input/inputCheckbox/styled/error.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import StyledError from '../../styled/error';
-
-const Error = styled(StyledError)`
-    align-self: flex-start;
-`;
-
-export default Error;
