--- src/components/Input/inputCheckbox/styled/helpText.js
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-import StyledHelpText from '../../styled/helpText';
-
-const HelpText = styled(StyledHelpText)`
-    align-self: flex-start;
-`;
-
-export default HelpText;
